+++ storage/pgsql/model/order.go
@@ -4,11 +4,11 @@ import "gophermart/model"
 
 // Order db model.
 type Order struct {
-	ID         string  `db:"id"`
-	UserID     uint64  `db:"user_id"`
-	Status     string  `db:"status"`
-	Accrual    float64 `db:"accrual"`
-	UploadedAT string  `db:"uploaded_at"`
+	ID         string `db:"id"`
+	UserID     uint64 `db:"user_id"`
+	Status     string `db:"status"`
+	Accrual    int    `db:"accrual"`
+	UploadedAT string `db:"uploaded_at"`
 }
 
 // ToCanonical converts db model to base model.
