+++ storage/pgsql/schema.go
@@ -5,15 +5,15 @@ create table if not exists users (
 	id bigserial not null,
 	login varchar (255) unique not null,
 	password varchar (255) not null,
-	balance decimal (16,2) not null default 0,
-	outcome decimal (16,2)  not null default 0
+	balance int not null default 0,
+	outcome int not null default 0
 );
 
 create table if not exists orders (
     id varchar (255) not null primary key,
     user_id integer not null,
     status varchar (50) not null default 'NEW',
-    accrual decimal (16,2) not null default 0,
+    accrual int not null default 0,
     uploaded_at timestamp(0) with time zone not null default current_timestamp
 );
 
@@ -21,7 +21,7 @@ create table if not exists transactions (
     id bigserial not null,
     user_id integer not null,
     order_id varchar (255) not null,
-    amount decimal (16, 2),
+    amount int,
     created_at timestamp(0) with time zone not null default current_timestamp
 );
 `
